tracepoint probes: filter by system and append to final probe
We 'turn on' the switch and implement the filter in
handle_query_module() which rejects any module that does not match the
given system (if any).

This commit also makes the system part of the final name of the
tracepoint. As a nice side-effect, listing tracepoints using e.g.

stap -l 'kernel.trace("*")'

will list tracepoints in order of system.