Allow syscall tests to match the arm64 remapped syscalls
The arm64 does not implement a number of deprecated syscalls.  The
deprecated syscalls are remapped to more generic syscalls with the
additional arguments set to so the remapped systemcall emulates the
behavior the deprecated syscalls.  The syscall tests matching needed
to be adjusted to match these remaped syscalls for inotify_init(),
pipe(), select(), and vfork().  These fixes are applied to the
following tests:

* systemtap.syscall/inotify.c
* systemtap.syscall/pipe.c
* systemtap.syscall/recv.c
* systemtap.syscall/recvfrom.c
* systemtap.syscall/recvmmsg.c
* stemtap.syscall/recvmsg.c
* systemtap.syscall/select.c
* systemtap.syscall/vforkwait.c