Split the access of sp from the rest of pt_regs regs[] accesses
The pt_regs struct has regs[0] through regs[30].  The sp is a separate
field and not regs[31] in the regs array.  The compiler will fail to
compile modules when the code includes accesses to the sp because of
the warnings.  This change splits out sp (and all the other reg)
accesses so this code compiles correctly without warnings.  This
issues observed with the systemtap.base/rlimit.exp tests failures.
The rlimit.exp tests now all pass with the patch.