Improve the output of cycle_thief.stp to include executable name
Jeremy Eder suggested that it would be good to include the executable
name in the output to save people the step of looking up the
executable name of the various PIDs.  The change also includes a
header for the output in that section to make it clearer what each
column is.