PR17737: tapset-reference html chapters should not get <toc>
The docbook xsl-stylesheets' html output normally includes generated
TOC for incoming <chapter> blocks.  Suppress this with a cunning
choice of generate.toc --stringparam for the xmlto.  (On F21 at least,
the parameter needs to contain an embedded \ apparently due to layers
of being passed between shell scripts and xml parsers and whatnot.
Sorry.)