Add "_best" category sample scripts
On suggestion from Brendan Gregg, it would be helpful for the examples
list to highlight "best" scripts, out of the over-one-hundred we ship.
To begin, this is approximated by special support for the .meta keyword
"_best", which results in highlighting the marked samples at the top
of the [keyword-]index.html files.