PR17858: Just warn and continue for stapdyn unknown globals
With staprun modules, the kernel just noted the unknown module option in
dmesg, and everything continues.  For stapdyn, we were getting a warning
which set enough of an error state to skip the rest of initialization,
including begin probes, but not enough to exit the script.  Now unknown
options just warn without setting error state, but invalid options like
bad number formats will set an error and properly exit.