tapset: tag many inline embedded-C blocks as pure
If embedded-C is not explicitly marked pure, then it is assumed to have
side effects, and can never be optimized away.  In many cases, we're
simply referencing a constant value, so this is certainly pure, and
should be tagged so.

This patch doesn't bother updating instances that are in impure contexts
anyway, like the many cases of @__syscall_gate et al. which test the
constant to perhaps issue probe "next".