session.cxx: improve parse error token colorization
Previous code had imperfect heuristics about cutting apart parts of
erroneous source lines, with a colorized token in the middle.  It
could occur that the token was synthetic, or cmd-line $/@ based,
in which case the srcline-cutting arithmetic could go off-end and
cause a basic_string::substr: exception.  New code dabbles less in
heuristics and confirms that the token content actually matches
the srcline content.  In case of a mismatch, the original srcline
around the token's location is transcribed verbatim (with ellipses).