new example: whythefail.stp
Inspired by wcohen's speculate.stp tapset/testcase and cibegttd@irc's
needs, this script prints a statement-execution trace for a given
function, but only for those runs of the function that ended up with a
(configurable) post-return condition.  It turns out to be a neat
multipurpose widget, for random-sampled function tracing as well as
failure analysis.