Use a value for bad syscall that works better on 32-bit ARM
On 32-bit ARM the use of -1 for the syscall number for the bad syscall
in the utrace_syscall_args test will result in a kernel oops on some
kernels and the use other negative numbers will result in illegal
instruction exceptions.  Changing the bad syscall to use a large
positive number for the bad syscall to allow this test to work better
on 32-bit arm.

or an illegal instruction fault.