PR18555: prefer linkage_name to match the symtab
DW_AT_name is usually only the same as the symbol table for C.  C++
names are mangled, which may be given by DW_AT_linkage_name.  So if we
want to compare a DWARF subprogram to the symbol table by name, we
should prefer the linkage name when it's available.

This mattered especially for ppc64le, where query_dwarf_func was trying
to apply the global/local symbol offset.  When we took a DWARF C++
function and tried to find that name in the symbol table for its offset,
there was no match, so the function wouldn't be resolved at all.

Now that lookup uses the linkage name.  If there's still no match, like
with a stripped symbol table, then it falls through to just use DWARF's
entrypc as usual.

This patch also maintains the raw "addr" and offset "entrypc" separately
for symbol table functions, so for instance update_symtab can still
compare the original address.