interned_string: construct from null char*
Some of the tapset*cxx functions were used to representing empty
function / file names with NULL char*s.  As these are transitioned to
interned_strings, the latter's ctors must be willing to accept those
NULLs and convert them to empty strings.  (Adjusting all the callers
to pass "" instead of NULL in those slots is also possible.)