runtime: only compare line records at "commit" states
The state machine for line records may go through intermediate states
which are not proper to compare address against.  Only at DW_LNS_copy,
DW_LNE_end_sequence, and special opcodes are we supposed to "append a
row to the matrix", as the DWARF specification tells us.

The usymfileline testcase didn't hit problems, because its C target only
outputted a simple sequence of special opcodes.  It's now C++, as that's
a very easy way to get code generated from a lot of different files and
lines.  A simple std::find creates pretty complicated line records, even
in an unoptimized build.