PR18936: extend the lifetime of synthesized DW_OP_addr
In dwflpp::literal_stmt_for_local(), a fake addr_loc is created for
variables we find in the symbol table.  But this addr_loc had tight
scope on the stack, and its contents were clobbered by the time the
pointer was read by emit_header().

Just extending the lifetime to the whole function scope is enough.