Hide "__global_" mangling from the module parameters
This mangling is an implementation detail of the module, and we don't
want to introduce a version dependency for staprun and stapdyn which
understand this.  Leave the parameter names unmangled and simply map
them to the mangled variable internally.

This also makes sure we don't expose the new private variables as
parameters at all, not even in mangled form.