Take a plain string for print_format::string_to_components
This is mostly called with print_format::raw_components, which is a string,
and an interned value isn't needed for sharing within, so avoid it.

It might be nicer to make raw_components an interned_string itself, but a
few places in tapsets.cxx are building that string incrementally, so they'd
need a bit of refactoring to use separate string buffers first.