Make the "shebang trick" work for RPM packaged testcases.
Some testcases use the "shebang trick" described in stap_run_batch()
testsuite function.  Commit d2995e6f breaks this for RPM packaged
testcases by putting the absolute path to stap into the shebang.

This update reverts cb69017 update to stap_run_batch() and fixes the
specfile shebang rewriter instead, so that it only applies to
systemtap examples contained in the systemtap-client package, where
the "shebang trick" is unwelcome.  With this update, testcases
contained in the systemtap-testsuite RPM remain unchanged, so that
the "shebang trick" works for them.