Subclass c_tmpcounter directly from c_unparser
The c_tmpcounter is meant to be a shadow of c_unparser, walking the same
walk to declare the temps that c_unparser will need.  But this was done
manually in a fragile balancing act, with a lot of duplicated logic and
little to ensure it was correct unless it failed to compile at all.

Now c_tmpcounter is a subclass of c_unparser so it can be a real shadow,
truly walking the same walk via the exact same code.  The normal output
is just dumped into a nulled ostream for this pass.  When a var is
accessed the first time through its value(), c_tmpcounter writes its
declaration to the real ostream.  A var that's created but not used will
be left out automatically, like a tmpvar that gets an override().