Eliminate the use of tex(nomencl.sty) in TeX documentation files
RHEL 7 does not package the nomencl.sty file and having this style
file used in systemtap makes it difficult to build the documentation.
The only place it was being used was in the tutorial.tex and it was
not generating a useful glossary.  The glossary generation of
tutorial.tex has been disabled for the time being.