Restrict optimization of literals in "%s\n" to streams only
This "use_print" optimization is only effective for printing to streams
with _stp_print().  Then a literal string argument to "%s\n" can just
have that literal token-pasted with the newline string.  But when
printing to strings, like sprintln, that optimization isn't actually
taken, so it was getting the newline twice.

Add new tests with sprint and sprintln on string literals, then clean up
the code that does the optimization -- for streams only.