Add a nop_visitor, and use it for expression_is_functioncall
In expression_is_functioncall, it only cares if the immediate expression
is a functioncall, but the expression_visitor it was derived from would
fully traverse all other expressions.  All it really wanted was like a
visitor-patterned dynamic_cast.

The new nop_visitor better enables this.  It does nothing at all for any
node types, and then expression_is_functioncall can override just for the
functioncall it cares about.

(We haven't yet settled whether to prefer a visitor pattern like this,
or more direct virtual calls like expression::is_symbol, or just plain
dynamic_cast<>.  Try to avoid the need for downcasts if possible.)