Restore the dedupe logic for tracepoint headers
With and without this dedupe logic, I see no difference in the
tracepoints found on Fedora 23, RHEL7, and RHEL6.  But it is faster
since we avoid processing headers found in both the kernel source tree
and build tree.

The way the header paths are normalized for #include, I don't see how
skipping that dedupe logic was supposed to help find more tracepoints.
I've added a comment pointing to what other functions would need to
change as well for that to have a more tangible effect, in case we
decide to compare file contents instead of just paths, for instance.

See also commit b1966849ecf that introduced the "#if 0" I'm removing.
The specific tracepoints mentioned in irq_vectors.h are still fine.