tapsets.cxx: fix indentation to appease GCC 6
The new -Wmisleading-indentation complains about the indentation of
these braces because of the if-statement right before. We can squash
that warning by simply unindenting the braces.

Whether this scoping should even exist is another question. I like it
there because that whole scope is related to variable expansion. It
should probably be moved to a function.