Make xgettext scan for all the NLS macros we use.
Before this update, xgettext, which is run as a part of `make update-po`
was only scanning for the '_()' macro (the other one mentioned in there,
'N_()', was not really used in our sources, it's probably some unused
relict). So now we scan for '_()', '_N()', '_F()', and '_NF()' as defined
in util.h.

From now on, the '_N()' and '_NF()' macros are correctly processing the
plural forms too. Relevant documentation is: https://www.gnu.org/softwa-
re/gettext/manual/html_node/xgettext-Invocation.html#Language-specific-options