Add ability to override kernel release check
In some situation it might be useful to disable checking the
kernel release. This happens when a kernel module is being rebuilt
and then probed. Without this override one has to reboot the machine
with the new kernel (and module) and then use systemtap.

To rectify the situation add a new define STP_NO_VERREL_CHECK,
which disables the relevant code in the resulting systemtap module.

This can be used with the following syntax:
stap -DSTP_NO_VERREL_CHECK script.stp

Signed-off-by: Nikolay Borisov <n.borisov@siteground.com>