Fix PR20211 by aimplementing the "resume mode" of the testsuite.
Once the SYSTEMTAP_TESTSUITE_RESUME environmental variable is set
to some non-zero value, the testsuite can run in "resume mode".

In this mode, tests, that have previously been run, are not being
re-executed, and their existing test results are being reused and
merged instead.

This can be useful e.g. when some of the tests causes kernel panic
or stall. Using the "resume mode", it is possible to reboot the
testing box, and re-run the testsuite skipping already executed
tests (incl. the one that caused the panic or stall - which will
stay "unfinished").

The "make list-unfinished" lists all such unfinished testcases.
Those need additional investigation.

The "resume mode" can be used with "make installcheck", or
"make installcheck-parallel".