java: add basic support for java objects
Previously stap was only able to probe primitive java types.

Instead of falling back on string cases if a primitive isn't found
(thus erroring out if the type doesn't match, such as an object), we
now explicitly check for the a string type, and default to converting
the object name to a string if no primitive is found and reporting that.

Originally reported: https://sourceware.org/ml/systemtap/2016-q4/msg00125.html