PR21020: reorganize data passing abi for java method parameters
Under a new ABI, all java object types are supported for passing to
java-method probes, by converting them all within the JVM to strings.
The systemtap script side now gets $argX being all utf-char-pointers,
which are loaded into script-side string variables with synthetic
user_string_warn() function calls.

A backward compatibility ABI is preserved in the java helper class
as well as the translator, under --compatible=3.0.

The java testsuite is drastically simplified (and execution time
shortened) to one class.  These now test string and object passing
too, with it and stapbm extended to be able to locate the HelperSDT
jar / .so files within the systemtap installation $prefix.  This way,
the system /usr/lib/jvm/.... path does not need to be populated just
for stap tests to run.  (We should sometime extend the tests to cover
all the argument arities, but eyeballing looks OK for now.)