PR20627: rework several code rewriting passes
The translator depends on some staptree rewriting passes not only for
optimization but also for correctness.

The @defined() construct is one where its timely execution and
implication handling is important, so that other constructs protected
by false-valued if(@defined()) guards are not even processed.  In
other places, @defined() nodes are collapsed to zero under heuristic
conditions.  This has all been a mash of uncomfortable probe-specific
and probe-independent code, running the same visitors in different
modes at different times, looping or not.

So we take a step forward in recognizing that some visitors
necessarily operate partially, and must be rerun, interleaved with
others.  New functions in tapsets.h:

update_visitor_loop() -- given a vector of update_visitors, loop till relaxed
var_expand_const_fold_loop() -- same specialized to const-folder

... are now used from general elaborator driver as well as
probe-derivation logic.  Some heuristic code regarding @defined()
can be removed.

This extends the update_visitor base template with its own notion of
relaxed-ness, which is naturally that if a staptree node/pointer has
been replaced (changed to a different value) during the visiting
traversal, then it is -not- relaxed yet, and this is saved in a new
instance variable.  (A reset() function resets this so the same
visitor object can be reused.)

This makes it unnecessary to track count_replaced_defined_ops()
separately.  It should make it possible to drop some other bool& flags
in some other older optimization passes; this is left for later.