symbol resolution: tweak levenshtein suggestions for private symbols
Attempted or typo'd access to a private function or variable properly
fail, but the levensthein suggestions should not include the
not-visible private symbols in the tapset.  This patch excludes the
private ones from suggestion lists.  But it is overbroad: a private
variable in the user script may also be excluded.