Add new probe alias input.char
input.char enables scripts to receive input from stdin during
module runtime. It is an alias for the probe procfs.write("__stdin").
The presence of this procfs probe puts the terminal in non-canonical mode
and causes a thread to spawn in staprun. This thread will read characters
from stdin one at a time and write them to /proc/systemtap/MODNAME/__stdin,
triggering the probe once per character.

An example of this functionality can be seen in EXAMPLES/general/eventcount_interactive.stp.
This script is a modified eventcount.stp where two of the display parameters can
be modified during runtime.