semantic analysis: functions: process them all in semantic_pass_symbols()
With a number of synthetic script functions being generated
during the probe-derivation process, we need to ensure that
all of them go through the full symbol-resolution process.
Previously, only the dome->functions[] array was dealt with
in this early part of pass-2, which was probably OK because
only embedded-C functions tended to be synthetically generated.
Not any more!