loc2c-runtime.h: in deref calls, tolerate nested calling
With the loc2stap code, it is possible to see constructs like
	deref(8, deref(8, deref(8, deref(8, fetch_register(4))))
so the addr field is itself a deref.  This blows up if the deref
macro copies the addr parameter multiple times, with a glorious
exponential fanout!  So now we use a temporary variable within
the macro, to hold the address value, then reuse that.