debuginfo diagnostics: complain about unresolved $vars better
dwarf_var_expanding_visitor::visit_target_symbol used to just quietly
sulk if a $context variable could not be resolved due to a null-die
reference scope.  This happens e.g. if no debuginfo was found for a
target binary, but the probe point was resolved perhaps due to symbol
tables.  It would result in a mystifying "unresolved target-symbol
expression" error.  The new message is thrown in context and treated
like a saved_conversion_error, with an explicit error::dwarf2 hint.