autocast support: bring back early manual functionall->synth-decl referent
When synthesizing dwarf-related functions, we can usually rely on
pass-2 symbol resolution to match up the functioncall*s with the
functiondecl*s.  But in the case of autocast handling, synth function
generation is for whatever reason delayed past this point in pass-2,
so we have to manually resolve symbols/referents to some extent.