bpf translation error handling: try tracking derived_probe token
BPF translation internals involve data structures that do not track
the staptree level entities they are based on.  So when an error
occurs, there's no quick way to mapping it back to the token, for an
ideal error message.  translate_bpf_pass() now tries to track a
last-token* as it goes translating all the probe bodies to bpf, so
that a semantic_error can be synthesized if needed.

This is still far from ideal; a single exception causes giving up
trying all the remaining probes; and all begin/end probes are
translated into single larger bpf programs, so it's harder to narrow
down the cause.  Baby steps.